style(msvc): fix C4244/C4267 in RageUtil/SongManager/XmlFileUtil/ThemeManager/ScreenEdit/RageFileManager (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226.

- RageUtil.cpp (25): SmEscape/DwiEscape/DelimitorLength/do_split's int
  len/size params and locals fed from RString::size()/pointer-diffs;
  two identical pcre_exec(..., sStr.size(), ...) calls (int length
  param); Trim*/Dirname's size-backed loop counters;
  Json::Value::resize (ArrayIndex, not size_t); a TONUMBER_NICE macro
  and multiapproach's mult local losing lua_tonumber's double into
  float; five lua_rawgeti/lua_rawseti/process-index sites in
  multiapproach's per-element loop.
- SongManager.cpp (20): SetTotalWork/wrap/RString::Left/Right
  size-into-int at load-time call sites; eight one-line GetNum*()
  getters returning size()/count_if() as int; three reverse-loop
  size()-1 bounds (safe by construction); a count_if accumulated into
  int iCount.
- XmlFileUtil.cpp (14): a local SetString(int,int,...) helper called
  from 6 sites with RString::size_type start/end positions -- cast at
  each call site rather than widening the helper's params, since its
  internal iEnd-1 >= iStart bounds check relies on signed underflow
  behavior that unsigned params would change; two reverse-loop bounds.
- ThemeManager.cpp (11): RString::Left/Right fed size_t positions;
  a reverse-loop bound and a size()-returning getter; two
  std::distance() results into int iDist; a lua_createtable/
  lua_rawseti pair pushing theme-fallback names.
- ScreenEdit.cpp (11): RandomInt(size_t); SCREEN_HEIGHT/480*0.5
  promoting to double via integer division then multiply, fixed with
  an outer static_cast<float> (x2); an int beats-per-measure value
  into a float; a choices.size()-1 default-choice index; a
  find()-begin() pointer-diff into unsigned pos plus a size() into a
  keysound-index int; a MenuRowDef size-derived count param; two
  size_t-indexed track-remap array writes; a size() into numKeysounds.
- RageFileManager.cpp (11): a Seek(int) fed an mz_uint64 file offset
  (the call already runtime-checks for truncation via a round-trip
  compare, so the cast doesn't change behavior); reverse-loop
  size()-1 bounds x3; RString::Left/Right fed size_t-derived
  mount-point-length arithmetic; three size()/length() values into int
  locals tracking driver/file counts.

+++ src/RageFileManager.cpp
@@ -119,7 +119,7 @@ std::size_t zipRead(void *pOpaque, mz_uint64 file_ofs, void *pBuf, std::size_t n
 {
 	RageFile *f = static_cast<RageFile*>(pOpaque);
 
-	const int pos = f->Seek(file_ofs);
+	const int pos = f->Seek(static_cast<int>(file_ofs));
 	if (pos >= 0 && static_cast<std::uint64_t>(pos) != file_ofs)
 	{
 		return 0;
@@ -466,7 +466,7 @@ RageFileManager::~RageFileManager()
 
 	/* Note that drivers can use previously-loaded drivers, eg. to load a ZIP
 	 * from the FS.  Unload drivers in reverse order. */
-	for( int i = g_pDrivers.size()-1; i >= 0; --i )
+	for( int i = static_cast<int>(g_pDrivers.size())-1; i >= 0; --i )
 	{
 		delete g_pDrivers[i]->m_pDriver;
 		delete g_pDrivers[i];
@@ -490,11 +490,11 @@ RString LoadedDriver::GetPath( const RString &sPath ) const
 			return RString();
 	}
 
-	if( sPath.Left(m_sMountPoint.size()).CompareNoCase(m_sMountPoint) )
+	if( sPath.Left(static_cast<int>(m_sMountPoint.size())).CompareNoCase(m_sMountPoint) )
 		return RString(); /* no match */
 
 	/* Add one, so we don't cut off the leading slash. */
-	RString sRet = sPath.Right( sPath.size() - m_sMountPoint.size() + 1 );
+	RString sRet = sPath.Right( static_cast<int>(sPath.size() - m_sMountPoint.size() + 1) );
 	return sRet;
 }
 
@@ -515,7 +515,7 @@ void RageFileManager::GetDirListing( const RString &sPath_, std::vector<RString>
 	ReferenceAllDrivers( apDriverList );
 
 	int iDriversThatReturnedFiles = 0;
-	int iOldSize = AddTo.size();
+	int iOldSize = static_cast<int>(AddTo.size());
 	for( unsigned i = 0; i < apDriverList.size(); ++i )
 	{
 		LoadedDriver *pLoadedDriver = apDriverList[i];
@@ -523,7 +523,7 @@ void RageFileManager::GetDirListing( const RString &sPath_, std::vector<RString>
 		if( p.empty() )
 			continue;
 
-		const unsigned OldStart = AddTo.size();
+		const unsigned OldStart = static_cast<unsigned>(AddTo.size());
 
 		pLoadedDriver->m_pDriver->GetDirListing( p, AddTo, bOnlyDirs, bReturnPathToo );
 		if( AddTo.size() != OldStart )
@@ -947,7 +947,7 @@ RString RageFileManager::ResolvePath(const RString &path)
 		if ( pDriver->m_sType != "dir" && pDriver->m_sType != "dirro" )
 			continue;
 
-		int iMountPointLen = pDriver->m_sMountPoint.length();
+		int iMountPointLen = static_cast<int>(pDriver->m_sMountPoint.length());
 		if( tmpPath.substr(0, iMountPointLen) != pDriver->m_sMountPoint )
 			continue;
 
@@ -1095,7 +1095,7 @@ RageFileBasic *RageFileManager::OpenForWriting( const RString &sPath, int mode, 
 	/* Only write files if they'll be read.  If a file exists in any driver, don't
 	 * create or write files in any driver mounted after it, because when we later
 	 * try to read it, we'll get that file and not the one we wrote. */
-	int iMaximumDriver = apDriverList.size();
+	int iMaximumDriver = static_cast<int>(apDriverList.size());
 	if( !Values.empty() && Values[0].second == 0 )
 		iMaximumDriver = Values[0].first;
 
@@ -1175,7 +1175,7 @@ void GetDirListingRecursive( const RString &sDir, const RString &sMatch, std::ve
 		vsDirs.erase( vsDirs.begin()+i );
 		i--;
 	}
-	for( int i=vsFiles.size()-1; i>=0; i-- )
+	for( int i=static_cast<int>(vsFiles.size())-1; i>=0; i-- )
 	{
 		if( !IsADirectory(vsFiles[i]) )
 			AddTo.push_back( vsFiles[i] );
@@ -1196,7 +1196,7 @@ void GetDirListingRecursive( RageFileDriver *prfd, const RString &sDir, const RS
 		vsDirs.erase( vsDirs.begin()+i );
 		i--;
 	}
-	for( int i=vsFiles.size()-1; i>=0; i-- )
+	for( int i=static_cast<int>(vsFiles.size())-1; i>=0; i-- )
 	{
 		if( prfd->GetFileType(vsFiles[i]) != RageFileManager::TYPE_DIR )
 			AddTo.push_back( vsFiles[i] );
